No longer uses tensorflow to load masks. Renamed job name in fsc_launch Moved open methods in process to utils.
